Refactor: Flatten modules and simplify structure
Flattened `src/ast/mod.rs` to `src/ast.rs` and `src/codegen/mod.rs` to `src/codegen.rs`.
Merged `src/errors/` directory into `src/errors.rs` to reduce nesting.
Merged `src/semantic/assembly_model.rs` into `src/semantic/assembler.rs`.
Merged `src/semantic/types.rs` into `src/semantic/model.rs`.

This adheres to the "Razor" philosophy of reducing unnecessary hierarchy and file count.
